refactor: extract chunking logic from index_vault.py into chunking.py
Moves 12 chunking functions and FRONTMATTER_EXCLUDE constant into
a dedicated src/chunking.py module, reducing blast radius of future
changes to either chunking or indexing orchestration.

Closes #46
